+++ src/Command/MakerCommand.php
@@ -15,7 +15,7 @@ use Symfony\Bundle\MakerBundle\ApplicationAwareMakerInterface;
 use Symfony\Bundle\MakerBundle\ConsoleStyle;
 use Symfony\Bundle\MakerBundle\DependencyBuilder;
 use Symfony\Bundle\MakerBundle\Exception\RuntimeCommandException;
-use Symfony\Bundle\MakerBundle\ExtraGenerationMakerInterface;
+use Symfony\Bundle\MakerBundle\FileManager;
 use Symfony\Bundle\MakerBundle\Generator;
 use Symfony\Bundle\MakerBundle\InputConfiguration;
 use Symfony\Bundle\MakerBundle\MakerInterface;
@@ -33,16 +33,16 @@ use Symfony\Component\Console\Output\OutputInterface;
 final class MakerCommand extends Command
 {
     private $maker;
-    private $generator;
+    private $fileManager;
     private $inputConfig;
     /** @var ConsoleStyle */
     private $io;
     private $checkDependencies = true;
 
-    public function __construct(MakerInterface $maker, Generator $generator)
+    public function __construct(MakerInterface $maker, FileManager $fileManager)
     {
         $this->maker = $maker;
-        $this->generator = $generator;
+        $this->fileManager = $fileManager;
         $this->inputConfig = new InputConfiguration();
 
         parent::__construct();
@@ -56,12 +56,14 @@ final class MakerCommand extends Command
     protected function initialize(InputInterface $input, OutputInterface $output)
     {
         $this->io = new ConsoleStyle($input, $output);
+        $this->fileManager->setIo($this->io);
 
         if ($this->checkDependencies) {
             $dependencies = new DependencyBuilder();
             $this->maker->configureDependencies($dependencies);
-            if ($missingPackages = $dependencies->getMissingDependencies()) {
-                throw new RuntimeCommandException(sprintf("Missing package%s: to use the %s command, run: \n\ncomposer require %s", 1 === count($missingPackages) ? '' : 's', $this->getName(), implode(' ', $missingPackages)));
+
+            if ($missingPackagesMessage = $dependencies->getMissingPackagesMessage($this->getName())) {
+                throw new RuntimeCommandException($missingPackagesMessage);
             }
         }
     }
@@ -86,15 +88,14 @@ final class MakerCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $this->generator->setIO($this->io);
-        $params = $this->maker->getParameters($input);
-        $this->generator->generate($params, $this->maker->getFiles($params));
+        $generator = new Generator($this->fileManager, 'App\\');
 
-        if ($this->maker instanceof ExtraGenerationMakerInterface) {
-            $this->maker->afterGenerate($this->io, $params);
+        $this->maker->generate($input, $this->io, $generator);
+
+        // sanity check for custom makers
+        if ($generator->hasPendingOperations()) {
+            throw new \LogicException('Make sure to call the writeChanges() method on the generator.');
         }
-
-        $this->maker->writeSuccessMessage($params, $this->io);
     }
 
     public function setApplication(Application $application = null)