+++ src/Maker/MakeController.php
@@ -15,14 +15,13 @@ use Doctrine\Common\Annotations\Annotation;
 use Symfony\Bundle\MakerBundle\ConsoleStyle;
 use Symfony\Bundle\MakerBundle\DependencyBuilder;
 use Symfony\Bundle\MakerBundle\FileManager;
+use Symfony\Bundle\MakerBundle\Generator;
 use Symfony\Bundle\MakerBundle\InputConfiguration;
 use Symfony\Bundle\MakerBundle\Str;
-use Symfony\Bundle\MakerBundle\Validator;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Routing\RouterInterface;
 
 /**
  * @author Javier Eguiluz <javier.eguiluz@gmail.com>
@@ -30,12 +29,10 @@ use Symfony\Component\Routing\RouterInterface;
  */
 final class MakeController extends AbstractMaker
 {
-    private $router;
     private $fileManager;
 
-    public function __construct(RouterInterface $router, FileManager $fileManager)
+    public function __construct(FileManager $fileManager)
     {
-        $this->router = $router;
         $this->fileManager = $fileManager;
     }
 
@@ -53,45 +50,40 @@ final class MakeController extends AbstractMaker
         ;
     }
 
-    public function getParameters(InputInterface $input): array
+    public function generate(InputInterface $input, ConsoleStyle $io, Generator $generator)
     {
-        $controllerClassName = Str::asClassName($input->getArgument('controller-class'), 'Controller');
-        Validator::validateClassName($controllerClassName);
-        $baseLayoutExists = $this->fileManager->fileExists('templates/base.html.twig');
+        $controllerClassNameDetails = $generator->createClassNameDetails(
+            $input->getArgument('controller-class'),
+            'Controller\\',
+            'Controller'
+        );
 
-        return [
-            'controller_class_name' => $controllerClassName,
-            'controller_class_file' => 'src/Controller/'.$controllerClassName.'.php',
-            'route_path' => Str::asRoutePath(str_replace('Controller', '', $controllerClassName)),
-            'route_name' => Str::asRouteName(str_replace('Controller', '', $controllerClassName)),
-            'twig_file' => Str::asFilePath(str_replace('Controller', '', $controllerClassName)).'.html.twig',
-            'base_layout_exists' => $baseLayoutExists,
-            'twig_installed' => $this->isTwigInstalled(),
-        ];
-    }
+        $templateName = Str::asFilePath($controllerClassNameDetails->getRelativeNameWithoutSuffix()).'/index.html.twig';
+        $controllerPath = $generator->generateClass(
+            $controllerClassNameDetails->getFullName(),
+            'controller/Controller.tpl.php',
+            [
+                'route_path' => Str::asRoutePath($controllerClassNameDetails->getRelativeNameWithoutSuffix()),
+                'route_name' => Str::asRouteName($controllerClassNameDetails->getRelativeNameWithoutSuffix()),
+                'twig_installed' => $this->isTwigInstalled(),
+                'template_name' => $templateName,
+            ]
+        );
 
-    public function getFiles(array $params): array
-    {
-        $dir = __DIR__.'/../Resources/skeleton/controller';
-
-        $paths = [$dir.'/Controller.tpl.php' => $params['controller_class_file']];
-
-        if ($params['twig_installed']) {
-            $paths[$dir.'/twig_template.tpl.php'] = 'templates/'.$params['twig_file'];
+        if ($this->isTwigInstalled()) {
+            $generator->generateFile(
+                'templates/'.$templateName,
+                'controller/twig_template.tpl.php',
+                [
+                    'base_layout_exists' => $this->fileManager->fileExists('templates/base.html.twig'),
+                    'controller_path' => $controllerPath,
+                ]
+            );
         }
 
-        return $paths;
-    }
+        $generator->writeChanges();
 
-    public function writeSuccessMessage(array $params, ConsoleStyle $io)
-    {
-        parent::writeSuccessMessage($params, $io);
-
-        if (!count($this->router->getRouteCollection())) {
-            $io->text('<error> Warning! </> No routes configuration defined yet.');
-            $io->text('           You should probably uncomment the annotation routes in <comment>config/routes.yaml</>');
-            $io->newLine();
-        }
+        $this->writeSuccessMessage($io);
         $io->text('Next: Open your new controller class and add some pages!');
     }