+++ src/Maker/MakeForm.php
@@ -11,39 +11,24 @@
 
 namespace Symfony\Bundle\MakerBundle\Maker;
 
-use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
 use Symfony\Bundle\MakerBundle\ConsoleStyle;
 use Symfony\Bundle\MakerBundle\DependencyBuilder;
-use Symfony\Bundle\MakerBundle\Doctrine\DoctrineEntityHelper;
-use Symfony\Bundle\MakerBundle\Exception\RuntimeCommandException;
-use Symfony\Bundle\MakerBundle\FileManager;
-use Symfony\Bundle\MakerBundle\GeneratorHelper;
+use Symfony\Bundle\MakerBundle\Generator;
 use Symfony\Bundle\MakerBundle\InputConfiguration;
 use Symfony\Bundle\MakerBundle\Str;
 use Symfony\Bundle\MakerBundle\Validator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Validator\Validation;
 
 /**
  * @author Javier Eguiluz <javier.eguiluz@gmail.com>
  * @author Ryan Weaver <weaverryan@gmail.com>
- * @author Sadicov Vladimir <sadikoff@gmail.com>
  */
 final class MakeForm extends AbstractMaker
 {
-    private $fileManager;
-    private $entityHelper;
-
-    public function __construct(FileManager $fileManager, DoctrineEntityHelper $entityHelper)
-    {
-        $this->fileManager = $fileManager;
-        $this->entityHelper = $entityHelper;
-    }
-
     public static function getCommandName(): string
     {
         return 'make:form';
@@ -54,82 +39,41 @@ final class MakeForm extends AbstractMaker
         $command
             ->setDescription('Creates a new form class')
             ->addArgument('name', InputArgument::OPTIONAL, sprintf('The name of the form class (e.g. <fg=yellow>%sType</>)', Str::asClassName(Str::getRandomTerm())))
-            ->addOption('entity', null, InputOption::VALUE_NONE, 'Generate form class from existing entity')
             ->setHelp(file_get_contents(__DIR__.'/../Resources/help/MakeForm.txt'))
         ;
     }
 
-    public function interact(InputInterface $input, ConsoleStyle $io, Command $command)
+    public function generate(InputInterface $input, ConsoleStyle $io, Generator $generator)
     {
-        $entityFormGeneration = $input->getOption('entity');
+        $formClassNameDetails = $generator->createClassNameDetails(
+            $input->getArgument('name'),
+            'Form\\',
+            'Type'
+        );
 
-        if ($this->entityHelper->isDoctrineConnected()) {
-            $entityClassName = Str::removeSuffix(Str::asClassName($input->getArgument('name')), 'Type');
+        $entityClassNameDetails = $generator->createClassNameDetails(
+            $formClassNameDetails->getRelativeNameWithoutSuffix(),
+            'Entity\\'
+        );
 
-            $entityExists = $this->fileManager->fileExists('src/Entity/'.$entityClassName.'.php');
+        $generator->generateClass(
+            $formClassNameDetails->getFullName(),
+            'form/Type.tpl.php',
+            [
+                'entity_class_exists' => class_exists($entityClassNameDetails->getFullName()),
+                'entity_full_class_name' => $entityClassNameDetails->getFullName(),
+                'entity_class_name' => $entityClassNameDetails->getShortName(),
+            ]
+        );
 
-            if ($entityFormGeneration && !$entityExists) {
-                throw new RuntimeCommandException(sprintf('Entity "%s" doesn\'t exists in your project. May be you would like to create it with "make:entity" command?', $entityClassName));
-            }
+        $generator->writeChanges();
 
-            if ((!$entityFormGeneration && !$entityExists) || $entityFormGeneration) {
-                return;
-            }
+        $this->writeSuccessMessage($io);
 
-            $io->block(
-                sprintf('Entity "%s" found in your project', $entityClassName),
-                null,
-                'fg=yellow'
-            );
-            $entityFormGeneration = $io->confirm('Do you want to generate a form for an entity?', true);
-
-            $input->setOption('entity', $entityFormGeneration);
-        } elseif ($entityFormGeneration) {
-            $io->block([
-                'Doctrine not found, falling back to simple form generation',
-            ], null, 'fg=yellow');
-            $input->setOption('entity', false);
-        }
-    }
-
-    public function getParameters(InputInterface $input): array
-    {
-        $formClassName = Str::asClassName($input->getArgument('name'), 'Type');
-        Validator::validateClassName($formClassName);
-
-        $entityFormGeneration = $input->getOption('entity');
-
-        $entityClassName = $entityFormGeneration ? Str::removeSuffix($formClassName, 'Type') : false;
-
-        $formFields = $entityFormGeneration ? $this->entityHelper->getFormFieldsFromEntity($entityClassName) : ['field_name'];
-
-        $helper = new GeneratorHelper();
-
-        return [
-            'helper' => $helper,
-            'form_class_name' => $formClassName,
-            'entity_class_name' => $entityClassName,
-            'form_fields' => $formFields,
-        ];
-    }
-
-    public function getFiles(array $params): array
-    {
-        return [
-            __DIR__.'/../Resources/skeleton/form/Type.tpl.php' => 'src/Form/'.$params['form_class_name'].'.php',
-        ];
-    }
-
-    public function writeSuccessMessage(array $params, ConsoleStyle $io)
-    {
-        parent::writeSuccessMessage($params, $io);
-
-        if (!$params['entity_class_name']) {
-            $io->text([
-                'Next: Add fields to your form and start using it.',
-                'Find the documentation at <fg=yellow>https://symfony.com/doc/current/forms.html</>',
-            ]);
-        }
+        $io->text([
+            'Next: Add fields to your form and start using it.',
+            'Find the documentation at <fg=yellow>https://symfony.com/doc/current/forms.html</>',
+        ]);
     }
 
     public function configureDependencies(DependencyBuilder $dependencies)
@@ -144,12 +88,6 @@ final class MakeForm extends AbstractMaker
             Validation::class,
             'validator',
             // add as an optional dependency: the user *probably* wants validation
-            false
-        );
-
-        $dependencies->addClassDependency(
-            DoctrineBundle::class,
-            'orm',
             false
         );
     }