+++ src/Resources/skeleton/controller/twig_template.tpl.php
@@ -19,7 +19,7 @@
 
     This friendly message is coming from:
     <ul>
-        <li>Your controller at <code><?= $controller_class_file; ?></code></li>
+        <li>Your controller at <code><?= $controller_path; ?></code></li>
         <li>Your template at <code><?= $relative_path ?></code></li>
     </ul>
 </div>