+++ src/Resources/skeleton/validator/Constraint.tpl.php
@@ -1,13 +1,13 @@
 <?= "<?php\n" ?>
 
-namespace App\Validator\Constraints;
+namespace <?= $namespace; ?>;
 
 use Symfony\Component\Validator\Constraint;
 
 /**
  * @Annotation
  */
-class <?= $constraint_class_name ?> extends Constraint
+class <?= $class_name ?> extends Constraint
 {
     /*
      * Any public properties become valid options for the annotation.